updated to use same training script in both backends
