+++ src/utils/asr_train_utils.py
@@ -1,0 +1,115 @@
+#!/usr/bin/env python
+
+# Copyright 2017 Johns Hopkins University (Shinji Watanabe)
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+
+import logging
+
+# chainer related
+import chainer
+from chainer import training
+
+
+class CompareValueTrigger(object):
+    '''Trigger invoked when key value getting bigger or lower than before
+
+    Args:
+        key (str): Key of value.
+        compare_fn: Function to compare the values.
+        trigger: Trigger that decide the comparison interval
+
+    '''
+
+    def __init__(self, key, compare_fn, trigger=(1, 'epoch')):
+        self._key = key
+        self._best_value = None
+        self._interval_trigger = training.util.get_trigger(trigger)
+        self._init_summary()
+        self._compare_fn = compare_fn
+
+    def __call__(self, trainer):
+        observation = trainer.observation
+        summary = self._summary
+        key = self._key
+        if key in observation:
+            summary.add({key: observation[key]})
+
+        if not self._interval_trigger(trainer):
+            return False
+
+        stats = summary.compute_mean()
+        value = float(stats[key])  # copy to CPU
+        self._init_summary()
+
+        if self._best_value is None:
+            # initialize best value
+            self._best_value = value
+            return False
+        elif self._compare_fn(self._best_value, value):
+            return True
+        else:
+            self._best_value = value
+            return False
+
+    def _init_summary(self):
+        self._summary = chainer.reporter.DictSummary()
+
+
+def make_batchset(data, batch_size, max_length_in, max_length_out, num_batches=0):
+    # sort it by input lengths (long to short)
+    sorted_data = sorted(data.items(), key=lambda data: int(
+        data[1]['ilen']), reverse=True)
+    logging.info('# utts: ' + str(len(sorted_data)))
+    # change batchsize depending on the input and output length
+    minibatch = []
+    start = 0
+    while True:
+        ilen = int(sorted_data[start][1]['ilen'])
+        olen = int(sorted_data[start][1]['olen'])
+        factor = max(int(ilen / max_length_in), int(olen / max_length_out))
+        # if ilen = 1000 and max_length_in = 800
+        # then b = batchsize / 2
+        # and max(1, .) avoids batchsize = 0
+        b = max(1, int(batch_size / (1 + factor)))
+        end = min(len(sorted_data), start + b)
+        minibatch.append(sorted_data[start:end])
+        if end == len(sorted_data):
+            break
+        start = end
+    if num_batches > 0:
+        minibatch = minibatch[:num_batches]
+    logging.info('# minibatches: ' + str(len(minibatch)))
+
+    return minibatch
+
+
+# TODO(watanabe) perform mean and variance normalization during the python program
+# and remove the data dump process in run.sh
+def converter_kaldi(batch, reader):
+    for data in batch:
+        feat = reader[data[0].encode('ascii', 'ignore')]
+        data[1]['feat'] = feat
+
+    return batch
+
+
+def delete_feat(batch):
+    for data in batch:
+        del data[1]['feat']
+
+    return batch
+
+
+def restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
+    '''Extension to restore snapshot'''
+    @training.make_extension(trigger=(1, 'epoch'))
+    def restore_snapshot(trainer):
+        _restore_snapshot(model, snapshot, load_fn)
+
+    return restore_snapshot
+
+
+def _restore_snapshot(model, snapshot, load_fn=chainer.serializers.load_npz):
+    load_fn(snapshot, model)
+    logging.info('restored from ' + str(snapshot))
